Use AbstractWriter's Metadata setter/accessor
By caching a local Metadata field in setMetadata but not passing it to
the abstract superclass, and not providing a getMetadata implementation,
the getMetadata method would always return null.

Instead we perform our validation in setMetadata and then defer to
super.setMetadata, so that getMetadata will just do the right thing.